# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Salary Manager",
        "description": "CRUD&Permisson+Role system.",
        "contact": {
            "email": "Hexagon14@mail.ru"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Main API Server"
        }
    ],
    "paths": {
        "/api/absence": {
            "get": {
                "tags": [
                    "Absences"
                ],
                "summary": "Get list of absences.",
                "description": "Returns list of absences.",
                "operationId": "getAbsenceList",
                "responses": {
                    "200": {
                        "description": "Returns list of absences."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Absences"
                ],
                "summary": "Creates absence.",
                "description": "Creates absence.",
                "operationId": "createAbsence",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "type": "2",
                                    "user_id": "4"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Returns absence."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    }
                ]
            }
        },
        "/api/absence/{id}": {
            "get": {
                "tags": [
                    "Absences"
                ],
                "summary": "Get one absence.",
                "description": "Returns one absence.",
                "operationId": "getAbsence",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Absence id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns absence."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Absences"
                ],
                "summary": "Deletes absence.",
                "description": "Deletes absence.",
                "operationId": "deleteAbsence",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Absence id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Absence deleted."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Absences"
                ],
                "summary": "Updates absence.",
                "description": "Updates absence parameters that were given in request.",
                "operationId": "updateAbsence",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Absence id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "type": "2",
                                    "user_id": "4"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns absence."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logs in user.",
                "description": "Returns token in responce.",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "login": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "login": "admin",
                                    "password": "admin"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns token."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logs user out.",
                "description": "Invalidates the token.",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Logs user out."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "Bearer Token": {
                "type": "apiKey",
                "description": "Laravel passport security.",
                "in": "header",
                "name": "Authorization",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/auth/login",
                        "tokenUrl": "http://localhost/auth/token",
                        "refreshUrl": "http://localhostauth/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}